pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sengen"
version = "0.1.0"
description = "Scenario Generator for training and evaluating LLM RL agents with focus on ethical decision making"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["llm", "rl", "ethics", "ai-safety"]
authors = [
    { name = "Caleb Maresca", email = "marescacc@gmail.com" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "gymnasium>=0.29.0",
    "langchain>=0.1.0",
    "langchain_openai",
    "langchain_community",
    "langchain_ollama>=0.1.0",
    "langchain_huggingface>=0.1.0",
    "transformers>=4.30.0",
    "torch>=2.0.0",
    "accelerate>=0.20.0",
    "bitsandbytes>=0.43.0",
    "pydantic>=2.0.0",
    "openai>=1.0.0",
    "numpy>=1.21.0",
    "pandas>=1.3.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=22.0.0",
    "isort>=5.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 88
target-version = ['py38']

[tool.isort]
profile = "black"
multi_line_output = 3 

[tool.hatch.build.targets.wheel]
packages = ["src/sengen"]